chore: restructuring, extract Alarm logic into /utils/alarm.cc and .h
